requirements.txt
Flask==3.0.3
google-api-python-client==2.136.0
google-auth-oauthlib==1.2.1
google-generativeai==0.8.0
PyPDF2==3.0.1
python-docx==1.1.2
